requirements.txt
icalendar
requests
python-dotenv

